+++ src/Controller.php
@@ -122,43 +122,31 @@ class Controller extends Speaker
     /**
      * Get attributes about the currently active track in the queue.
      *
-     * @return array Track data containing the following elements (title, artist, album, track-number, queue-number, duration, position, stream)
+     * @return State Track data containing the following elements
      */
     public function getStateDetails()
     {
         $data = $this->soap("AVTransport", "GetPositionInfo");
 
         if (!$data["TrackMetaData"]) {
-            return [
-                "title"         =>  "",
-                "artist"        =>  "",
-                "album"         =>  "",
-                "track-number"  =>  0,
-                "album-art"     =>  "",
-                "queue-number"  =>  0,
-                "duration"      =>  $data["TrackDuration"],
-                "position"      =>  $data["RelTime"],
-                "stream"        =>  false,
-            ];
+            return new State;
         }
 
         $parser = new XmlParser($data["TrackMetaData"]);
-        $meta = Helper::getTrackMetaData($parser);
+        $state = State::createFromXml($parser, $this);
 
-        $stream = false;
         if ((string) $parser->getTag("streamContent")) {
             $info = $this->soap("AVTransport", "GetMediaInfo");
-            if (!$stream = (string) (new XmlParser($info["CurrentURIMetaData"]))->getTag("title")) {
-                $stream = (string) $parser->getTag("title");
+            if (!$state->stream = (string) (new XmlParser($info["CurrentURIMetaData"]))->getTag("title")) {
+                $state->stream = (string) $parser->getTag("title");
             }
         }
 
-        return array_merge($meta, [
-            "queue-number"  =>  (int) $data["Track"],
-            "duration"      =>  $data["TrackDuration"],
-            "position"      =>  $data["RelTime"],
-            "stream"        =>  $stream,
-        ]);
+        $state->queueNumber = (int) $data["Track"];
+        $state->duration = $data["TrackDuration"];
+        $state->position = $data["RelTime"];
+
+        return $state;
     }