+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Changelog
 ### Changed
 
 * [Network] The Network class is no longer static, it should be instantiated before calling its methods.
+* [Controllers] Controller::getStateDetails() nows return an instance of the State class
+* [Queues/Playlists] Queue::getTracks() now returns an array of QueueTrack instances
 * [Playlists] Creating playlists is now done using the createPlaylist() method on the Network class
 
 --------